+++ src/oldnews/providers/main.py
@@ -33,6 +33,7 @@ from ..commands import (
     RefreshFromTheOldReader,
     Remove,
     Rename,
+    ToggleCompact,
     ToggleShowAll,
     UserInformation,
 )
@@ -71,6 +72,7 @@ class MainCommands(CommandsProvider):
         yield RefreshFromTheOldReader()
         yield from self.maybe(Rename)
         yield from self.maybe(Remove)
+        yield ToggleCompact()
         yield ToggleShowAll()
         yield from self.maybe(UserInformation)