+++ src/oldnews/widgets/__init__.py
@@ -4,10 +4,18 @@
 # Local imports.
 from .article_content import ArticleContent
 from .article_list import ArticleList
+from .article_list_header import ArticleListHeader
+from .article_view import ArticleView
 from .navigation import Navigation
 
 ##############################################################################
 # Exports.
-__all__ = ["ArticleContent", "ArticleList", "Navigation"]
+__all__ = [
+    "ArticleContent",
+    "ArticleList",
+    "ArticleListHeader",
+    "ArticleView",
+    "Navigation",
+]
 
 ### __init__.py ends here